- calculate volume usd only once - moved poolVolume into volume helper

+++ src/utils/Token.ts
@@ -43,9 +43,9 @@ export function createAndReturnToken(tokenAddress: Address, converterAddress: Ad
   let liquidityPoolToken = LiquidityPoolToken.load(converterAddress.toHex() + tokenAddress.toHex())
   if (liquidityPoolToken == null) {
     liquidityPoolToken = new LiquidityPoolToken(converterAddress.toHex() + tokenAddress.toHex())
-    liquidityPoolToken.totalVolume = BigInt.zero()
-    liquidityPoolToken.volumeBought = BigInt.zero()
-    liquidityPoolToken.volumeSold = BigInt.zero()
+    liquidityPoolToken.totalVolume = BigDecimal.zero()
+    liquidityPoolToken.volumeBought = BigDecimal.zero()
+    liquidityPoolToken.volumeSold = BigDecimal.zero()
   }
 
   liquidityPoolToken.token = tokenAddress.toHex()
